ai_module/requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
openai==1.10.0
python-dotenv==1.0.0
nltk==3.8.1
supabase==2.0.0
langchain-openai==0.0.5

